requirements.txt
pydantic
fastapi
uvicorn
sqlalchemy
passlib
bcrypt==4.0.1
PyJWT
python-dotenv
python-multipart